requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
Flask-Session==0.5.0
Werkzeug==2.3.7
